feat: migrate to nativewindui components
