+++ packages/native-ui/package.json
@@ -7,7 +7,6 @@
   "exports": {
     "./blocks/*": "./src/blocks/*.tsx",
     "./components/*": "./src/components/*.tsx",
-    "./_components/*": "./src/_components/*.tsx",
     "./globals.css": "./src/styles/globals.css",
     "./hooks/*": "./src/hooks/*.ts",
     "./icons/*": "./src/icons/*.ts",
@@ -26,33 +25,13 @@
   "dependencies": {
     "@init/utils": "workspace:*",
     "@react-navigation/native": "catalog:",
-    "@rn-primitives/accordion": "1.1.0",
-    "@rn-primitives/alert-dialog": "^1.1.0",
-    "@rn-primitives/aspect-ratio": "^1.1.0",
-    "@rn-primitives/avatar": "^1.1.0",
-    "@rn-primitives/checkbox": "^1.1.0",
-    "@rn-primitives/collapsible": "1.1.0",
-    "@rn-primitives/context-menu": "^1.1.0",
-    "@rn-primitives/dialog": "^1.1.0",
-    "@rn-primitives/dropdown-menu": "^1.1.0",
+    "@rn-primitives/avatar": "1.1.0",
     "@rn-primitives/hooks": "^1.2.0",
-    "@rn-primitives/hover-card": "^1.1.0",
-    "@rn-primitives/label": "^1.1.0",
-    "@rn-primitives/menubar": "^1.1.0",
-    "@rn-primitives/navigation-menu": "^1.1.0",
-    "@rn-primitives/popover": "^1.1.0",
-    "@rn-primitives/progress": "^1.1.0",
-    "@rn-primitives/radio-group": "^1.1.0",
-    "@rn-primitives/select": "^1.1.0",
-    "@rn-primitives/separator": "^1.1.0",
     "@rn-primitives/slot": "1.1.0",
-    "@rn-primitives/switch": "^1.1.0",
-    "@rn-primitives/table": "^1.1.0",
     "@rn-primitives/tabs": "^1.1.0",
     "@rn-primitives/toggle": "^1.1.0",
     "@rn-primitives/toggle-group": "^1.1.0",
     "@rn-primitives/tooltip": "^1.1.0",
-    "@rn-primitives/types": "1.1.0",
     "@roninoss/icons": "catalog:",
     "@tanstack/react-form": "catalog:",
     "class-variance-authority": "catalog:",
