--- packages/native-ui/src/components/aspect-ratio.tsx
@@ -1,5 +1,0 @@
-import * as AspectRatioPrimitive from "@rn-primitives/aspect-ratio"
-
-const AspectRatio = AspectRatioPrimitive.Root
-
-export { AspectRatio }
