--- packages/native-ui/src/components/collapsible.tsx
@@ -1,9 +1,0 @@
-import * as CollapsiblePrimitive from "@rn-primitives/collapsible"
-
-const Collapsible = CollapsiblePrimitive.Root
-
-const CollapsibleTrigger = CollapsiblePrimitive.Trigger
-
-const CollapsibleContent = CollapsiblePrimitive.Content
-
-export { Collapsible, CollapsibleTrigger, CollapsibleContent }
