--- packages/native-ui/src/components/dropdown-menu.tsx
@@ -1,261 +1,0 @@
-import * as DropdownMenuPrimitive from "@rn-primitives/dropdown-menu"
-import * as React from "react"
-import {
-  Platform,
-  type StyleProp,
-  StyleSheet,
-  Text,
-  type TextProps,
-  View,
-  type ViewStyle,
-} from "react-native"
-
-import { cn } from "@init/utils/ui"
-
-import { Check, ChevronDown, ChevronRight, ChevronUp } from "./icon"
-import { TextClassContext } from "./text"
-
-const DropdownMenu = DropdownMenuPrimitive.Root
-
-const DropdownMenuTrigger = DropdownMenuPrimitive.Trigger
-
-const DropdownMenuGroup = DropdownMenuPrimitive.Group
-
-const DropdownMenuPortal = DropdownMenuPrimitive.Portal
-
-const DropdownMenuSub = DropdownMenuPrimitive.Sub
-
-const DropdownMenuRadioGroup = DropdownMenuPrimitive.RadioGroup
-
-const DropdownMenuSubTrigger = React.forwardRef<
-  DropdownMenuPrimitive.SubTriggerRef,
-  DropdownMenuPrimitive.SubTriggerProps & {
-    inset?: boolean
-  }
->(({ className, inset, children, ...props }, ref) => {
-  const { open } = DropdownMenuPrimitive.useSubContext()
-  const Icon =
-    Platform.OS === "web" ? ChevronRight : open ? ChevronUp : ChevronDown
-  return (
-    <TextClassContext.Provider
-      value={cn(
-        "select-none text-sm native:text-lg text-primary",
-        open && "native:text-accent-foreground"
-      )}
-    >
-      <DropdownMenuPrimitive.SubTrigger
-        ref={ref}
-        className={cn(
-          "flex web:cursor-default web:select-none flex-row items-center gap-2 rounded-sm px-2 native:py-2 py-1.5 web:outline-none web:hover:bg-accent web:focus:bg-accent active:bg-accent",
-          open && "bg-accent",
-          inset && "pl-8",
-          className
-        )}
-        {...props}
-      >
-        {children}
-        <Icon size={18} className="ml-auto text-foreground" />
-      </DropdownMenuPrimitive.SubTrigger>
-    </TextClassContext.Provider>
-  )
-})
-DropdownMenuSubTrigger.displayName =
-  DropdownMenuPrimitive.SubTrigger.displayName
-
-const DropdownMenuSubContent = React.forwardRef<
-  DropdownMenuPrimitive.SubContentRef,
-  DropdownMenuPrimitive.SubContentProps
->(({ className, ...props }, ref) => {
-  const { open } = DropdownMenuPrimitive.useSubContext()
-  return (
-    <DropdownMenuPrimitive.SubContent
-      ref={ref}
-      className={cn(
-        "data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-right-2 data-[side=right]:slide-in-from-left-2 data-[side=top]:slide-in-from-bottom-2 z-50 mt-1 min-w-[8rem] overflow-hidden rounded-md border border-border bg-popover p-1 shadow-foreground/5 shadow-md",
-        open
-          ? "web:fade-in-0 web:zoom-in-95 web:animate-in"
-          : "web:fade-out-0 web:zoom-out web:animate-out",
-        className
-      )}
-      {...props}
-    />
-  )
-})
-DropdownMenuSubContent.displayName =
-  DropdownMenuPrimitive.SubContent.displayName
-
-const DropdownMenuContent = React.forwardRef<
-  DropdownMenuPrimitive.ContentRef,
-  DropdownMenuPrimitive.ContentProps & {
-    overlayStyle?: StyleProp<ViewStyle>
-    overlayClassName?: string
-    portalHost?: string
-  }
->(
-  (
-    { className, overlayClassName, overlayStyle, portalHost, ...props },
-    ref
-  ) => {
-    const { open } = DropdownMenuPrimitive.useRootContext()
-    return (
-      <DropdownMenuPrimitive.Portal hostName={portalHost}>
-        <DropdownMenuPrimitive.Overlay
-          style={
-            overlayStyle
-              ? StyleSheet.flatten([
-                  Platform.OS !== "web" ? StyleSheet.absoluteFill : undefined,
-                  overlayStyle,
-                ])
-              : Platform.OS !== "web"
-                ? StyleSheet.absoluteFill
-                : undefined
-          }
-          className={overlayClassName}
-        >
-          <DropdownMenuPrimitive.Content
-            ref={ref}
-            className={cn(
-              "web:data-[side=bottom]:slide-in-from-top-2 web:data-[side=left]:slide-in-from-right-2 web:data-[side=right]:slide-in-from-left-2 web:data-[side=top]:slide-in-from-bottom-2 z-50 min-w-[8rem] overflow-hidden rounded-md border border-border bg-popover p-1 shadow-foreground/5 shadow-md",
-              open
-                ? "web:fade-in-0 web:zoom-in-95 web:animate-in"
-                : "web:fade-out-0 web:zoom-out-95 web:animate-out",
-              className
-            )}
-            {...props}
-          />
-        </DropdownMenuPrimitive.Overlay>
-      </DropdownMenuPrimitive.Portal>
-    )
-  }
-)
-DropdownMenuContent.displayName = DropdownMenuPrimitive.Content.displayName
-
-const DropdownMenuItem = React.forwardRef<
-  DropdownMenuPrimitive.ItemRef,
-  DropdownMenuPrimitive.ItemProps & {
-    inset?: boolean
-  }
->(({ className, inset, ...props }, ref) => (
-  <TextClassContext.Provider value="select-none text-sm native:text-lg text-popover-foreground web:group-focus:text-accent-foreground">
-    <DropdownMenuPrimitive.Item
-      ref={ref}
-      className={cn(
-        "group relative flex web:cursor-default flex-row items-center gap-2 rounded-sm px-2 native:py-2 py-1.5 web:outline-none web:hover:bg-accent web:focus:bg-accent active:bg-accent",
-        inset && "pl-8",
-        props.disabled && "web:pointer-events-none opacity-50",
-        className
-      )}
-      {...props}
-    />
-  </TextClassContext.Provider>
-))
-DropdownMenuItem.displayName = DropdownMenuPrimitive.Item.displayName
-
-const DropdownMenuCheckboxItem = React.forwardRef<
-  DropdownMenuPrimitive.CheckboxItemRef,
-  DropdownMenuPrimitive.CheckboxItemProps
->(({ className, children, checked, ...props }, ref) => (
-  <DropdownMenuPrimitive.CheckboxItem
-    ref={ref}
-    className={cn(
-      "web:group relative flex web:cursor-default flex-row items-center rounded-sm native:py-2 py-1.5 pr-2 pl-8 web:outline-none web:focus:bg-accent active:bg-accent",
-      props.disabled && "web:pointer-events-none opacity-50",
-      className
-    )}
-    checked={checked}
-    {...props}
-  >
-    <View className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
-      <DropdownMenuPrimitive.ItemIndicator>
-        <Check size={14} strokeWidth={3} className="text-foreground" />
-      </DropdownMenuPrimitive.ItemIndicator>
-    </View>
-    {children}
-  </DropdownMenuPrimitive.CheckboxItem>
-))
-DropdownMenuCheckboxItem.displayName =
-  DropdownMenuPrimitive.CheckboxItem.displayName
-
-const DropdownMenuRadioItem = React.forwardRef<
-  DropdownMenuPrimitive.RadioItemRef,
-  DropdownMenuPrimitive.RadioItemProps
->(({ className, children, ...props }, ref) => (
-  <DropdownMenuPrimitive.RadioItem
-    ref={ref}
-    className={cn(
-      "web:group relative flex web:cursor-default flex-row items-center rounded-sm native:py-2 py-1.5 pr-2 pl-8 web:outline-none web:focus:bg-accent active:bg-accent",
-      props.disabled && "web:pointer-events-none opacity-50",
-      className
-    )}
-    {...props}
-  >
-    <View className="absolute left-2 flex h-3.5 w-3.5 items-center justify-center">
-      <DropdownMenuPrimitive.ItemIndicator>
-        <View className="h-2 w-2 rounded-full bg-foreground" />
-      </DropdownMenuPrimitive.ItemIndicator>
-    </View>
-    {children}
-  </DropdownMenuPrimitive.RadioItem>
-))
-DropdownMenuRadioItem.displayName = DropdownMenuPrimitive.RadioItem.displayName
-
-const DropdownMenuLabel = React.forwardRef<
-  DropdownMenuPrimitive.LabelRef,
-  DropdownMenuPrimitive.LabelProps & {
-    inset?: boolean
-  }
->(({ className, inset, ...props }, ref) => (
-  <DropdownMenuPrimitive.Label
-    ref={ref}
-    className={cn(
-      "web:cursor-default px-2 py-1.5 font-semibold native:text-base text-foreground text-sm",
-      inset && "pl-8",
-      className
-    )}
-    {...props}
-  />
-))
-DropdownMenuLabel.displayName = DropdownMenuPrimitive.Label.displayName
-
-const DropdownMenuSeparator = React.forwardRef<
-  DropdownMenuPrimitive.SeparatorRef,
-  DropdownMenuPrimitive.SeparatorProps
->(({ className, ...props }, ref) => (
-  <DropdownMenuPrimitive.Separator
-    ref={ref}
-    className={cn("-mx-1 my-1 h-px bg-border", className)}
-    {...props}
-  />
-))
-DropdownMenuSeparator.displayName = DropdownMenuPrimitive.Separator.displayName
-
-const DropdownMenuShortcut = ({ className, ...props }: TextProps) => {
-  return (
-    <Text
-      className={cn(
-        "ml-auto native:text-sm text-muted-foreground text-xs tracking-widest",
-        className
-      )}
-      {...props}
-    />
-  )
-}
-DropdownMenuShortcut.displayName = "DropdownMenuShortcut"
-
-export {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuGroup,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuPortal,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuShortcut,
-  DropdownMenuSub,
-  DropdownMenuSubContent,
-  DropdownMenuSubTrigger,
-  DropdownMenuTrigger,
-}
