+++ packages/utils/package.json
@@ -24,7 +24,6 @@
     "nanoid": "5.1.3",
     "schema-dts": "1.1.5",
     "tailwind-merge": "3.0.2",
-    "tailwind-merge2": "npm:tailwind-merge@2.6.0",
     "zod": "catalog:",
     "zod-form-data": "2.0.7"
   },
